Cleaned Example app
