+++ Example/SheetKit/Utils/ColorUtils.swift
@@ -16,7 +16,7 @@ extension UIColor {
             return .black
         }
     }
-    
+
     class var flatRed: UIColor { return UIColor(hexValue: 0xc0392b) }
     class var flatYellow: UIColor { return UIColor(hexValue: 0xf1c40f) }
     class var flatOrange: UIColor { return UIColor(hexValue: 0xe67e22) }
@@ -24,29 +24,16 @@ extension UIColor {
     class var flatGray: UIColor { return UIColor(hexValue: 0x7f8c8d) }
     class var flatRock: UIColor { return UIColor(hexValue: 0x2c3e50) }
     class var flatGreen: UIColor { return UIColor(hexValue: 0x27ae60) }
-    
-    convenience init(red: Float, green: Float, blue: Float) {
+
+    convenience init(hexValue: Int) {
+        let red = (hexValue >> 16) & 0xFF
+        let green = (hexValue >> 8) & 0xFF
+        let blue = hexValue & 0xFF
+        
         assert(red >= 0 && red <= 255, "Invalid red component")
         assert(green >= 0 && green <= 255, "Invalid green component")
         assert(blue >= 0 && blue <= 255, "Invalid blue component")
         
-        self.init(
-            displayP3Red: CGFloat(red) / 255.0,
-            green: CGFloat(green) / 255.0,
-            blue: CGFloat(blue) / 255.0,
-            alpha: 1.0
-        )
-    }
-    
-    convenience init(red: Int, green: Int, blue: Int) {
-        self.init(red: Float(red), green: Float(green), blue: Float(blue))
-    }
-    
-    convenience init(hexValue: Int) {
-        self.init(
-            red: (hexValue >> 16) & 0xFF,
-            green: (hexValue >> 8) & 0xFF,
-            blue: hexValue & 0xFF
-        )
+        self.init(displayP3Red: CGFloat(red)/255, green: CGFloat(green)/255, blue: CGFloat(blue)/255, alpha: 1)
     }
 }
